fix UI loading box

+++ Teek_AR/Assets/ResponseModels/EventModel.cs
@@ -24,7 +24,7 @@ namespace Assets.ResponseModels
         public virtual Nullable<double> Price { get; set; }
         public virtual Nullable<int> Status { get; set; }
         public virtual bool EnableInteractionPage { get; set; }
-        public virtual Nullable<int> Multiplier { get; set; }
+        public virtual Nullable<double> Multiplier { get; set; }
         public virtual bool Active { get; set; }
         public virtual string MasterCode { get; set; }
         public List<ActivityModel> Activities { get; set; }
